# server/docs/swagger.yaml
basePath: /
definitions:
  handlers.AccountUpdateRequest:
    properties:
      currency:
        example: EUR
        type: string
      email:
        example: alice@example.com
        type: string
      username:
        example: alice
        type: string
    type: object
  handlers.AuthResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user: {}
    type: object
  handlers.BrandRequest:
    properties:
      empty_spool_weight_grams:
        example: 250
        type: number
      name:
        example: Bambu Lab
        type: string
      website:
        example: https://bambulab.com
        type: string
    type: object
...
  handlers.LoginRequest:
    properties:
      email:
        example: alice@example.com
        type: string
      password:
        example: s3cr3t!
        type: string
    type: object
  handlers.PasswordUpdateRequest:
    properties:
      current_password:
        example: old_pass
        type: string
      new_password:
        example: new_pass
        type: string
    type: object
...
  handlers.RegisterRequest:
    properties:
      email:
        example: alice@example.com
        type: string
      password:
        example: s3cr3t!
        type: string
      username:
        example: alice
        type: string
    type: object
...
  models.Brand:
    properties:
      created_at:
        type: string
      empty_spool_weight_grams:
        type: number
      id_brand:
        type: string
      id_origin:
        type: string
      name:
        type: string
      website:
        type: string
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      currency:
        type: string
      email:
        type: string
      id_user:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Filament inventory management for 3D printing enthusiasts.
  title: Filaventory API
  version: "1.0"
paths:
  /api/account:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get current user account
      tags:
      - account
    put:
      consumes:
      - application/json
      parameters:
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.AccountUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: unauthorized
          schema:
            type: string
        "409":
          description: email or username already in use
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update account details
      tags:
      - account
  /api/account/password:
    put:
      consumes:
      - application/json
      parameters:
      - description: Current and new password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.PasswordUpdateRequest'
      responses:
        "204":
          description: No Content
        "401":
          description: incorrect current password
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - account
  /api/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AuthResponse'
        "401":
          description: invalid credentials
          schema:
            type: string
      summary: Login
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.AuthResponse'
        "400":
          description: validation error
          schema:
            type: string
        "409":
          description: email or username already exists
          schema:
            type: string
      summary: Register a new user
      tags:
      - auth
  /api/brands:
    get:
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/models.Brand'
                type: array
              page:
                type: integer
              per_page:
                type: integer
              total:
                type: integer
            type: object
      security:
      - BearerAuth: []
      summary: List brands
      tags:
      - brands
    post:
      consumes:
      - application/json
      parameters:
      - description: Brand data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.BrandRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Brand'
        "400":
          description: name is required
          schema:
            type: string
        "409":
          description: brand already exists
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a brand
      tags:
      - brands
  /api/brands/{id}:
    delete:
      parameters:
      - description: Brand UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a brand
      tags:
      - brands
    get:
      parameters:
      - description: Brand UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Brand'
        "404":
          description: not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get a brand
      tags:
      - brands
    put:
      consumes:
      - application/json
      parameters:
      - description: Brand UUID
        in: path
        name: id
        required: true
        type: string
      - description: Brand data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.BrandRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Brand'
        "404":
          description: not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update a brand
      tags:
      - brands
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer {token}" in the Authorization header
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"